+++ client/src/api/photographyPackage.api.ts
@@ -1,6 +1,7 @@
 import axios, { AxiosResponse } from 'axios';
 import { domain } from '../config';
 import { PhotographyPackage } from '../interfaces/photographyPackage.interface';
+import { isTokenValid } from '../utils/checkToken';
 
 export const getPhotographyPackages = async (): Promise<PhotographyPackage[]> => {
     try {
@@ -12,9 +13,11 @@ export const getPhotographyPackages = async (): Promise<PhotographyPackage[]> =>
     }
 };
 
-export const addPhotographyPackage = async (newPackage: PhotographyPackage): Promise<PhotographyPackage> => {
+export const addPhotographyPackage = async (newPackage: PhotographyPackage) => {
     try {
-        const token: string | null = sessionStorage.getItem('token');
+        const token: string | null = isTokenValid();
+        if (!isTokenValid()) { return; }
+
         const response: AxiosResponse<PhotographyPackage> = await axios.post(`${domain}/PhotographyPackage`, newPackage, {
             headers: {
                 'token': token,
@@ -28,9 +31,11 @@ export const addPhotographyPackage = async (newPackage: PhotographyPackage): Pro
     }
 };
 
-export const updatePhotographyPackage = async (id: number, updatedPackage: Partial<PhotographyPackage>): Promise<PhotographyPackage> => {
+export const updatePhotographyPackage = async (id: number, updatedPackage: Partial<PhotographyPackage>) => {
     try {
-        const token: string | null = sessionStorage.getItem('token');
+        const token: string | null = isTokenValid();
+        if (!isTokenValid()) { return; }
+
         const response: AxiosResponse<PhotographyPackage> = await axios.put(`${domain}/PhotographyPackage/${id}`, updatedPackage, {
             headers: {
                 'token': token,
@@ -44,9 +49,11 @@ export const updatePhotographyPackage = async (id: number, updatedPackage: Parti
     }
 };
 
-export const deletePhotographyPackage = async (id: number): Promise<void> => {
+export const deletePhotographyPackage = async (id: number) => {
     try {
-        const token: string | null = sessionStorage.getItem('token');
+        const token: string | null = isTokenValid();
+        if (!isTokenValid()) { return; }
+
         await axios.delete(`${domain}/PhotographyPackage/${id}`, {
             headers: {
                 'token': token
